move fixture
from tests/conftest.py to tests/test-create-newsletter-subscriber.py
