+++ tests/conftest.py
@@ -67,18 +67,3 @@ class AuthActions(object):
 @pytest.fixture
 def auth(client):
     return AuthActions(client)
-
-
-class Newsletter(object):
-    def __init__(self, client):
-        self._client = client
-
-    def signup(self, email='test'):
-        return self._client.post(
-            '/newsletter/signup',
-            data={'email': email}
-        )
-
-
-def newsletter(client):
-    return Newsletter(client)
